backend/menuSystem/requirements.txt
Django==3.2.12
djangorestframework==3.15.1
django-cors-headers==4.3.1